+++ src/student_success_tool/modeling/training_h2o.py
@@ -1,22 +1,25 @@
-import mlflow
-from mlflow.tracking import MlflowClient
-
-import os
-import datetime
-import contextlib
 import logging
 import typing as t
-import tempfile
 
+from mlflow.tracking import MlflowClient
 import pandas as pd
 from pandas.api.types import is_categorical_dtype, is_object_dtype, is_string_dtype
-
-from . import evaluation_h2o as evaluation
 
 import h2o
 from h2o.automl import H2OAutoML
 
+from . import utils_h2o as utils
+
 LOGGER = logging.getLogger(__name__)
+
+VALID_H2O_METRICS = {
+    "auc",
+    "logloss",
+    "mean_per_class_error",
+    "rmse",
+    "mae",
+    "aucpr",
+}
 
 
 def run_h2o_automl_classification(
@@ -26,6 +29,7 @@ def run_h2o_automl_classification(
     primary_metric: str,
     institution_id: str,
     student_id_col: str,
+    client: t.Optional["MLflowClient"] = None,
     **kwargs: object,
 ) -> tuple[H2OAutoML, h2o.H2OFrame, h2o.H2OFrame, h2o.H2OFrame]:
     """
@@ -43,27 +47,8 @@ def run_h2o_automl_classification(
         Trained H2OAutoML object.
     """
 
-    # Validate input types
-    if not isinstance(df, pd.DataFrame):
-        raise TypeError("df must be a pandas DataFrame.")
-    if not isinstance(target_col, str):
-        raise TypeError("target_col must be a string.")
-    if target_col not in df.columns:
-        raise ValueError(f"target_col '{target_col}' not found in input DataFrame.")
-
-    VALID_H2O_METRICS = {
-        "auc",
-        "logloss",
-        "mean_per_class_error",
-        "rmse",
-        "mae",
-        "aucpr",
-    }
-    primary_metric = primary_metric.lower()
-    if primary_metric not in VALID_H2O_METRICS:
-        raise ValueError(
-            f"Invalid primary_metric '{primary_metric}'. Must be one of {VALID_H2O_METRICS}"
-        )
+    if client is None:
+        client = MlflowClient()
 
     # Set defaults and pop kwargs
     seed = kwargs.pop("seed", 42)
@@ -72,6 +57,36 @@ def run_h2o_automl_classification(
     exclude_cols = kwargs.pop("exclude_cols", [])
     split_col = kwargs.pop("split_col", "split")
     sample_weights_col = kwargs.pop("sample_weights_col", None)
+    target_name = kwargs.pop("target_name", None)
+    checkpoint_name = kwargs.pop("checkpoint_name", None)
+    workspace_path = kwargs.pop("workspace_path", None)
+    primary_metric = primary_metric.lower()
+
+    # Validate inputs
+    if target_col not in df.columns:
+        raise ValueError(f"target_col '{target_col}' not found in input DataFrame.")
+
+    if primary_metric not in VALID_H2O_METRICS:
+        raise ValueError(
+            f"Invalid primary_metric '{primary_metric}'. Must be one of {VALID_H2O_METRICS}"
+        )
+
+    missing_params = [
+        name
+        for name, value in {
+            "target_name": target_name,
+            "checkpoint_name": checkpoint_name,
+            "workspace_path": workspace_path,
+        }.items()
+        if not value
+    ]
+    if missing_params:
+        raise ValueError(f"Missing logging parameters: {', '.join(missing_params)}")
+
+    if split_col not in df.columns:
+        raise ValueError(
+            "Input data must contain a 'split' column with values ['train', 'validate', 'test']."
+        )
 
     if student_id_col and student_id_col not in exclude_cols:
         exclude_cols.append(student_id_col)
@@ -81,16 +96,13 @@ def run_h2o_automl_classification(
     h2o_df = h2o.H2OFrame(df)
     h2o_df = correct_h2o_dtypes(h2o_df, df)
 
-    if split_col not in h2o_df.columns:
-        raise ValueError(
-            "Input data must contain a 'split' column with values ['train', 'validate', 'test']."
-        )
-
+    # Convert dtype of target_col for h2o training, and create splits
     h2o_df[target_col] = h2o_df[target_col].asfactor()
     train = h2o_df[h2o_df[split_col] == "train"]
     valid = h2o_df[h2o_df[split_col] == "validate"]
     test = h2o_df[h2o_df[split_col] == "test"]
 
+    # Define feature list
     features = [col for col in df.columns if col not in exclude_cols + [target_col]]
 
     LOGGER.info(
@@ -116,287 +128,20 @@ def run_h2o_automl_classification(
 
     LOGGER.info(f"Best model: {aml.leader.model_id}")
 
-    return aml, train, valid, test
-
-
-def log_h2o_experiment(
-    aml: H2OAutoML,
-    *,
-    train: h2o.H2OFrame,
-    valid: h2o.H2OFrame,
-    test: h2o.H2OFrame,
-    institution_id: str,
-    target_col: str,
-    target_name: str,
-    checkpoint_name: str,
-    workspace_path: str,
-    client: t.Optional["MLflowClient"] = None,
-) -> tuple[str, pd.DataFrame]:
-    """
-    Logs evaluation metrics, plots, and model artifacts for all models in an H2O AutoML leaderboard to MLflow.
-
-    Args:
-        aml: Trained H2OAutoML object.
-        train: H2OFrame containing the training split.
-        valid: H2OFrame containing the validation split.
-        test: H2OFrame containing the test split.
-        institution_id: Institution identifier, used to namespace the MLflow experiment.
-        target_col: Column name of target (used for plotting and label extraction).
-        target_name: Name of the target of the model from the config.
-        checkpoint_name: Name of the checkpoint of the model from the config.
-        workspace_path: Path prefix for experiment naming within MLflow.
-        client: Optional MLflowClient instance. If not provided, one will be created.
-
-    Returns:
-        experiment_id (str): The MLflow experiment ID used for logging.
-        results_df (pd.DataFrame): DataFrame with metrics and MLflow run IDs for all successfully logged models.
-    """
-    LOGGER.info("Logging experiment to MLflow with classification plots...")
-
-    if client is None:
-        client = MlflowClient()
-
-    experiment_id = set_or_create_experiment(
-        workspace_path,
-        institution_id,
-        target_name,
-        checkpoint_name,
+    utils.log_h2o_experiment(
+        aml=aml,
+        train=train,
+        valid=valid,
+        test=test,
+        institution_id=institution_id,
+        target_col=target_col,
+        target_name=target_name,
+        checkpoint_name=checkpoint_name,
+        workspace_path=workspace_path,
         client=client,
     )
 
-    leaderboard_df = aml.leaderboard.as_data_frame()
-
-    log_h2o_experiment_summary(
-        aml=aml,
-        leaderboard_df=leaderboard_df,
-        train=train,
-        target_col=target_col,
-    )
-
-    # Limiting # of models that we're logging to save some time
-    MAX_MODELS_TO_LOG = 50
-    top_model_ids = leaderboard_df["model_id"].tolist()[:MAX_MODELS_TO_LOG]
-
-    if not top_model_ids:
-        LOGGER.warning("No models found in leaderboard.")
-        return experiment_id, pd.DataFrame()
-
-    results = []
-    num_models = len(top_model_ids)
-
-    for idx, model_id in enumerate(top_model_ids):
-        # Show status update
-        model_num = idx + 1
-
-        if model_num == 1 or model_num % 10 == 0 or model_num == num_models:
-            LOGGER.info(
-                f"Evaluating model {model_num}/{len(top_model_ids)}: {model_id}"
-            )
-
-        # Setting threshold to 0.5 due to binary classification
-        metrics = evaluate_and_log_model(
-            aml=aml,
-            model_id=model_id,
-            train=train,
-            valid=valid,
-            test=test,
-            threshold=0.5,
-            client=client,
-        )
-
-        if metrics:
-            results.append(metrics)
-
-    results_df = pd.DataFrame(results)
-    LOGGER.info(f"Logged {len(results_df)} model runs to MLflow.")
-
-    return experiment_id, results_df
-
-
-def evaluate_and_log_model(
-    aml: H2OAutoML,
-    *,
-    model_id: str,
-    train: h2o.H2OFrame,
-    valid: h2o.H2OFrame,
-    test: h2o.H2OFrame,
-    client: "MLflowClient",
-    threshold: float = 0.5,
-) -> dict | None:
-    """
-    Evaluates a single H2O model at a given threshold and logs metrics, plots, and artifacts to MLflow.
-
-    Args:
-        model_id (str): The H2O model ID to evaluate and log.
-        aml: H2OAutoML object containing the leaderboard and trained models.
-        train: H2OFrame with training data.
-        valid: H2OFrame with validation data.
-        test: H2OFrame with test data.
-        threshold (float): Threshold to apply for binary classification metrics.
-        client (MlflowClient): Initialized MLflow client used for logging.
-
-    Returns:
-        dict: Dictionary of evaluation metrics including `mlflow_run_id`, or `None` if evaluation fails.
-    """
-    try:
-        model = h2o.get_model(model_id)
-
-        with (
-            open(os.devnull, "w") as fnull,
-            contextlib.redirect_stdout(fnull),
-            contextlib.redirect_stderr(fnull),
-        ):
-            metrics = evaluation.get_metrics_near_threshold_all_splits(
-                model, train, valid, test, threshold=threshold
-            )
-
-            if mlflow.active_run():
-                mlflow.end_run()
-
-            with mlflow.start_run(run_name=f"h2o_eval_{model_id}"):
-                run_id = mlflow.active_run().info.run_id
-
-                for k, v in metrics.items():
-                    if k != "model_id":
-                        mlflow.log_metric(k, v)
-
-                for split_name, frame in zip(
-                    ["train", "val", "test"], [train, valid, test]
-                ):
-                    y_true = frame["target"].as_data_frame().values.flatten()
-                    preds = model.predict(frame)
-                    positive_class_label = preds.col_names[-1]
-                    y_proba = (
-                        preds[positive_class_label].as_data_frame().values.flatten()
-                    )
-                    y_pred = (y_proba >= 0.5).astype(int)
-
-                    evaluation.generate_all_classification_plots(
-                        y_true, y_pred, y_proba, prefix=split_name
-                    )
-
-                local_model_dir = f"/tmp/h2o_models/{model_id}"
-                os.makedirs(local_model_dir, exist_ok=True)
-                h2o.save_model(model, path=local_model_dir, force=True)
-                mlflow.log_artifacts(local_model_dir, artifact_path="model")
-
-        metrics["mlflow_run_id"] = run_id
-        return metrics
-
-    except Exception as e:
-        LOGGER.exception(f"Failed to log model {model_id}: {e}")
-        return None
-
-
-def log_h2o_experiment_summary(
-    *,
-    aml: H2OAutoML,
-    leaderboard_df: pd.DataFrame,
-    train: h2o.H2OFrame,
-    target_col: str,
-    run_name: str = "h2o_automl_experiment_summary",
-    sample_size: int = 1000,
-) -> None:
-    """
-    Logs summary information about the H2O AutoML experiment to a dedicated MLflow run in
-    the experiment with the leaderboard as a CSV, list of input features, training dataset,
-    target distribution, and the schema (column names and types)
-
-    Args:
-        aml: Trained H2OAutoML object.
-        leaderboard_df (pd.DataFrame): Leaderboard as DataFrame.
-        train (H2OFrame): Training H2OFrame.
-        target_col (str): Name of the target column.
-        run_name (str): Name of the MLflow run. Defaults to "h2o_automl_experiment_summary".
-        sample_size (int): Number of rows to sample from train data. Defaults to 1000.
-    """
-    if mlflow.active_run():
-        mlflow.end_run()
-
-    with mlflow.start_run(run_name=run_name):
-        # Log basic experiment metadata
-        mlflow.log_metric("num_models_trained", len(leaderboard_df))
-        mlflow.log_param("best_model_id", aml.leader.model_id)
-
-        # Create tmp directory for artifacts
-        with tempfile.TemporaryDirectory() as tmpdir:
-            # Log leaderboard
-            leaderboard_path = os.path.join(tmpdir, "h2o_leaderboard.csv")
-            leaderboard_df.to_csv(leaderboard_path, index=False)
-            mlflow.log_artifact(leaderboard_path, artifact_path="leaderboard")
-
-            # Log feature list
-            features = [col for col in train.columns if col != target_col]
-            features_path = os.path.join(tmpdir, "train_features.txt")
-            with open(features_path, "w") as f:
-                for feat in features:
-                    f.write(f"{feat}\n")
-            mlflow.log_artifact(features_path, artifact_path="inputs")
-
-            # Log sampled training data
-            train_df = train.as_data_frame(use_pandas=True)
-            train_path = os.path.join(tmpdir, "train.csv")
-            train_df.to_csv(train_path, index=False)
-            mlflow.log_artifact(train_path, artifact_path="inputs")
-
-            # Log target distribution
-            target_dist_df = train[target_col].table().as_data_frame()
-            target_dist_path = os.path.join(tmpdir, "target_distribution.csv")
-            target_dist_df.to_csv(target_dist_path, index=False)
-            mlflow.log_artifact(target_dist_path, artifact_path="inputs")
-
-            # Log schema
-            schema_df = pd.DataFrame(train.types.items(), columns=["column", "dtype"])
-            schema_path = os.path.join(tmpdir, "train_schema.csv")
-            schema_df.to_csv(schema_path, index=False)
-            mlflow.log_artifact(schema_path, artifact_path="inputs")
-
-
-def set_or_create_experiment(
-    workspace_path: str,
-    institution_id: str,
-    target_name: str,
-    checkpoint_name: str,
-    client: t.Optional["MlflowClient"] = None,
-) -> str:
-    """
-    Creates or retrieves a structured MLflow experiment and sets it as the active experiment.
-
-    Args:
-        workspace_path: Base MLflow workspace path.
-        institution_id: Institution or tenant identifier used for experiment naming.
-        target_name: Name of the target variable.
-        checkpoint_name: Name of the modeling checkpoint.
-        client: MLflow client. A new one is created if not provided.
-
-    Returns:
-        MLflow experiment ID (created or retrieved).
-    """
-    if client is None:
-        client = MlflowClient()
-
-    timestamp = datetime.datetime.utcnow().strftime("%Y-%m-%dT%H:%M:%S")
-
-    name_parts = [institution_id, target_name, checkpoint_name, "h20_automl", timestamp]
-    experiment_name = "/".join(
-        [
-            workspace_path.rstrip("/"),
-            "h2o_automl",
-            "_".join([part for part in name_parts if part]),
-        ]
-    )
-
-    try:
-        experiment = client.get_experiment_by_name(experiment_name)
-        if experiment is None:
-            experiment_id = client.create_experiment(experiment_name)
-        else:
-            experiment_id = experiment.experiment_id
-
-        mlflow.set_experiment(experiment_name)
-        return experiment_id
-    except Exception as e:
-        raise RuntimeError(f"Failed to create or set MLflow experiment: {e}")
+    return aml, train, valid, test
 
 
 def correct_h2o_dtypes(
